pyproject.toml
[tool.poetry]
name = "bpm"
version = "0.1.0"
description = ""
authors = ["Azat Seidakhmet <seidakhmetazat@gmail.com>"]
readme = "README.md"
packages = [
    { include = "apps", from = "src" },
    { include = "config", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2"
black = "^23.12.1"
ipython = "7.34"
psycopg2-binary = "^2.9.9"
environ = "^1.0"
django-environ = "^0.11.2"
pycryptodome = "^3.20.0"
requests = "^2.31.0"
django-object-actions = "^4.2.0"
djangorestframework = "^3.14.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
